requirements.txt
asgiref==3.8.0
Django==4.2.2
djangorestframework==3.15.0
pillow==10.2.0
sqlparse==0.4.4
tzdata==2024.1
